added api folders and moved controler routes
