+++ controllers/index.js
@@ -1,15 +1,10 @@
 const router = require('express').Router();
 
 const homeRoutes = require('./homeRoutes');
-const orderRoutes = require('./orderRoutes');
-const userRoutes = require('./userRoutes');
-const productRoutes = require('./productRoutes');
-const orderDetailRoutes = require('./orderDetailRoutes')
+const apiRoutes = require('./api');
+
 
 
 router.use('/', homeRoutes);
-router.use('/order', orderRoutes);
-router.use('/user', userRoutes);
-router.use('/product', productRoutes);
-router.use('/order_detail', orderDetailRoutes);
+router.use('/api', apiRoutes);
 module.exports = router;
